core: rename folders, format code
