+++ src/pages/dashboard/video-games/descuentos/components/DescuentoColumnHeaders.tsx
